Add casts to eliminate warnings when sizeof(int) is 16-bits
git-svn-id: svn://svn.code.sf.net/p/nuttx/code/trunk@2401 42af7a65-404d-4744-a932-0658087f49c3

+++ drivers/mmcsd/mmcsd_csd.h
@@ -283,7 +283,7 @@
 /* C_SIZE 62-73 Device size */
 
 #define MMCSD_CSD_CSIZE(csd) (((csd[6] & 3) << 10) | (csd[7] << 2) | (csd[8] >> 6))
-#define SD20_CSD_CSIZE(csd)  (((csd[7] & 0x3f) << 16) | (csd[8] << 8) | csd[9])
+#define SD20_CSD_CSIZE(csd)  ((((uint32_t)csd[7] & 0x3f) << 16) | (csd[8] << 8) | csd[9])
 
 /* VDD_R_CURR_MIN 59-61 = Max. read current at Vcc min */
